requirements.txt
absl-py==2.1.0
adjustText==1.3.0
ag2==0.7.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
astunparse==1.6.3
asyncer==0.0.8
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
codecarbon==3.0.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
et_xmlfile==2.0.0
fast-depends==2.4.12
fastjsonschema==2.21.2
fief-client==0.20.0
fix-busted-json==0.0.18
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.58.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
keras==3.9.0
kiwisolver==1.4.8
Levenshtein==0.27.1
libclang==18.1.1
logparser3==1.0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.4.1
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.0
nvidia-ml-py==12.570.86
ollama==0.4.7
openai==1.60.1
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
pandas==2.2.3
pandocfilters==1.5.1
pillow==11.2.1
prometheus_client==0.21.1
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyautogen==0.7.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pynvml==12.0.0
pyparsing==3.2.3
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2024.2
questionary==2.1.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2022.3.2
requests==2.32.3
rich==13.9.4
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tqdm==4.67.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarg==0.1.9
yaspin==3.1.0
